% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gg_12_1}
\alias{gg_12_1}
\title{Greengenes training set 12_1}
\format{
A data frame with 3 columns. Each row represents a different
sequence:
\describe{
\item{id}{Sequence accession identifier}
\item{sequence}{DNA sequence string}
\item{taxonomy}{Taxonomic string with each level separated with a \verb{;}}
}
}
\source{
\itemize{
\item \href{https://mothur.org/wiki/greengenes-formatted_databases/}{mothur-formatted files}
\item \href{https://pubmed.ncbi.nlm.nih.gov/21716311/}{Werner et al.}
}
}
\usage{
gg_12_1
}
\description{
The sequence and taxonomy data for the full-length sequences found in
Greengene's SSU (16S/18S) database version 12_1 for use with the naive
Bayesian classifier as implemented in the \code{{phylyotypr}} R package. This
includes 84,414 bacterial and archaeal sequences. Originally released by the
Greengenes website in January 2012 (now defunct) and published by Werner et al.
See the pages in the "Sources" section for more information.
}
\keyword{datasets}
