% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gg_13_5}
\alias{gg_13_5}
\title{Greengenes training set 13_8}
\format{
A data frame with 3 columns. Each row represents a different
sequence:
\describe{
\item{id}{Sequence accession identifier}
\item{sequence}{DNA sequence string}
\item{taxonomy}{Taxonomic string with each level separated with a \verb{;}}
}

A data frame with 3 columns. Each row represents a different
sequence:
\describe{
\item{id}{Sequence accession identifier}
\item{sequence}{DNA sequence string}
\item{taxonomy}{Taxonomic string with each level separated with a \verb{;}}
}
}
\source{
\itemize{
\item \href{https://mothur.org/wiki/greengenes-formatted_databases/}{mothur-formatted files}
}

\itemize{
\item \href{https://mothur.org/wiki/greengenes-formatted_databases/}{mothur-formatted files}
}
}
\usage{
gg_13_5

gg_13_5
}
\description{
The sequence and taxonomy data for the full-length sequences found in
Greengene's SSU (16S/18S) database version 13_8 for use with the naive
Bayesian classifier as implemented in the \code{{phylyotypr}} R package. This
includes 329,175 bacterial and 2,094 archaeal sequences. Originally released
by the biocore group in May 2013. See the pages in the "Sources" section
for more information.

The sequence and taxonomy data for the full-length sequences found in
Greengene's SSU (16S/18S) database version 13_8 for use with the naive
Bayesian classifier as implemented in the \code{{phylyotypr}} R package. This
includes 202,421 bacterial and archaeal sequences. Originally released by the
Greengenes website in May 2013 (now defunct). See the pages in the "Sources"
section for more information.
}
\keyword{datasets}
