% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gg_20_10}
\alias{gg_20_10}
\title{Greengenes training set 20_10}
\format{
A data frame with 3 columns. Each row represents a different
sequence:
\describe{
\item{id}{Sequence accession identifier}
\item{sequence}{DNA sequence string}
\item{taxonomy}{Taxonomic string with each level separated with a \verb{;}}
}

An object of class \code{data.frame} with 331269 rows and 3 columns.
}
\source{
\itemize{
\item \href{https://mothur.org/wiki/greengenes2-formatted_databases/}{mothur-formatted files}
\item \href{https://mothur.org/blog/2024/greengenes2_2020_10/}{Description of how mothur-formatted files were generated}
\item \href{https://github.com/biocore/greengenes2/}{Biocore's release notes for this version}
}
}
\usage{
gg_20_10

gg_20_10
}
\description{
The sequence and taxonomy data for the full-length sequences found in
Greengene's SSU (16S/18S) database version 20_10 for use with the naive
Bayesian classifier as implemented in the \code{{phylyotypr}} R package. This
includes 329,175 bacterial and 2,094 archaeal sequences. Originally released
by the biocore group in October 2022. Note that the biocore version includes
species names, but these have been removed from the mothur version of the
database. See the pages in the "Sources" section for more information.
}
\keyword{datasets}
