% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdp_trainset16}
\alias{rdp_trainset16}
\alias{rdp_trainset16_pds}
\title{RDP training set v16}
\format{
A data frame with 3 columns. Each row represents a different
sequence:
\describe{
\item{id}{Sequence accession identifier}
\item{sequence}{DNA sequence string}
\item{taxonomy}{Taxonomic string with each level separated with a \verb{;}}
}

An object of class \code{data.frame} with 13335 rows and 3 columns.
}
\source{
\itemize{
\item \href{https://mothur.org/wiki/rdp_reference_files/}{mothur-formatted files}
\item \href{https://mothur.org/blog/2017/RDP-v16-reference_files/}{Description of how mothur-formatted files were generated}
\item \href{https://sourceforge.net/projects/rdp-classifier/files/RDP_Classifier_TrainingData/}{RDP sourceforge page}
}
}
\usage{
rdp_trainset16

rdp_trainset16_pds
}
\description{
The sequence and taxonomy data for the 13,212 sequences found in the
Ribosomal Database Project's trainset16_022016 training set for use with the
naive Bayesian classifier as implemented in the \code{{phylyotypr}} R package.
Originally released by the RDP in February 2016. The \code{rdp} version contains
}
\keyword{datasets}
