% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rdp_trainset18}
\alias{rdp_trainset18}
\alias{rdp_trainset18_pds}
\title{RDP training set v18}
\format{
A data frame with 3 columns. Each row represents a different
sequence:
\describe{
\item{id}{Sequence accession identifier}
\item{sequence}{DNA sequence string}
\item{taxonomy}{Taxonomic string with each level separated with a \verb{;}}
}

An object of class \code{data.frame} with 21318 rows and 3 columns.
}
\source{
\itemize{
\item \href{https://mothur.org/wiki/rdp_reference_files/}{mothur-formatted files}
\item \href{https://mothur.org/blog/2021/RDP-v18-reference-files/}{Description of how mothur-formatted files were generated}
\item \href{https://sourceforge.net/projects/rdp-classifier/files/RDP_Classifier_TrainingData/}{RDP sourceforge page}
}
}
\usage{
rdp_trainset18

rdp_trainset18_pds
}
\description{
The sequence and taxonomy data for the 21,313 sequences found in the
Ribosomal Database Project's trainset18_062020 training set for use with the
naive Bayesian classifier as implemented in the \code{{phylyotypr}} R package.
Originally released by the RDP in June 2020. The \code{rdp} version contains the
same sequences as provided by the official RDP version. The \code{pds} version
contains 119 eukaryotic sequences including chloroplasts and mitochondria.
See the mothur reference file page in "Sources" for more information.
}
\keyword{datasets}
