% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{silva_v138_2_nr}
\alias{silva_v138_2_nr}
\alias{silva_v138_2_seed}
\title{Silva training set v138.2}
\format{
A data frame with 3 columns. Each row represents a different
sequence:
\describe{
\item{id}{Sequence accession identifier}
\item{sequence}{DNA sequence string}
\item{taxonomy}{Taxonomic string with each level separated with a \verb{;}}
}

An object of class \code{data.frame} with 8696 rows and 3 columns.
}
\source{
\itemize{
\item \href{https://mothur.org/wiki/silva_reference_files/}{mothur-formatted files}
\item \href{https://mothur.org/blog/2024/SILVA-v138_2-reference-files/}{Description of how mothur-formatted files were generated}
\item \href{https://www.arb-silva.de/documentation/release-1382/}{Silva release notes for this version}
}
}
\usage{
silva_v138_2_nr

silva_v138_2_seed
}
\description{
The sequence and taxonomy data for the full-length sequences found in Silva's
SSU (16S/18S) database version 138.2 for use with the naive Bayesian
classifier as implemented in the \code{{phylyotypr}} R package. This includes
sequences from bacteria, archaea, and eukaryotes. Originally released by
Silva in July 2024. The \code{nr} version contains the non-redundant version of
the database, which excludes sequences more than 99\% similar to other
sequences (164,296 total sequences); some additional polishing of the
sequences and taxonomy have been applied. The \code{seed} version contains 8,707
sequences that are thought to be part of their SEED database for purposes of
aligning. The sequences provided here are not aligned. See the mothur
reference file page in "Sources" for more information.
}
\keyword{datasets}
